+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 2.12.0
 
 * Added a view for test cases. This is planned to evolve into a suite management feature over the next releases
+* Fixed the order of display of quick-jump items, and they are now sorted by name
 
 ## Version 2.11.1